requirements.txt
absl-py==2.1.0
accelerate==0.22.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aioice==0.9.0
aiortc==1.7.0
aiosignal==1.3.1
albumentations==1.3.1
altair==4.2.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
basicsr==1.4.2
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blinker==1.7.0
braceexpand==0.1.7
Brotli==1.1.0
brotlipy==0.7.0
cachetools==5.3.2
certifi==2020.12.5
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1613413863336/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1610093492116/work
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
coloredlogs==15.0.1
conda==4.10.1
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1618231390031/work
contourpy==1.2.0
cryptography==42.0.2
customtkinter==5.1.3
cycler==0.12.1
Cython==3.0.8
darkdetect==0.8.0
decorator==4.4.2
decord==0.6.0
diffusers==0.21.4
dlib==19.24.2
dnspython==2.5.0
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.11
einops==0.6.1
entrypoints==0.4
exceptiongroup==1.2.0
facenet-pytorch==2.5.3
facexlib==0.2.5
fastapi==0.109.2
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
filterpy==1.4.5
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
gast==0.4.0
gdown==5.0.0
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.41
glfw==2.7.0
google-auth==2.25.2
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-pasta==0.2.0
gradio==3.41.2
gradio-client==0.5.0
grpcio==1.60.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.1.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
ifaddr==0.2.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
importlib-resources==6.1.1
jax==0.4.23
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langdetect==1.0.9
lazy-loader==0.3
libclang==16.0.6
lightning-utilities==0.10.1
lit==17.0.6
llvmlite==0.41.1
lmdb==1.4.1
loguru==0.7.2
lpips==0.1.4
mamba @ file:///home/conda/feedstock_root/build_artifacts/mamba_1620856587040/work
mangum==0.17.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
natsort==8.4.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onedrivedownloader==1.1.3
onnxruntime==1.16.3
opencv-python==4.8.0.76
opencv-python-headless==4.6.0.66
opennsfw2==0.10.2
opt-einsum==3.3.0
orjson==3.9.13
packaging==23.2
pandas==2.2.1
pillow==10.3.0
platformdirs==4.1.0
prettytable==3.9.0
proglog==0.1.10
protobuf==4.23.2
psutil==5.9.5
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyBanubaSDK @ file:///home/ubuntu/PyBanubaSDK-1.12.0/PyBanubaSDK.1.12.0.cpython-39-x86_64-linux-gnu
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1610094799048/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic-core==2.16.3
pydeck==0.8.1b0
pyDeprecate==0.3.1
pydub==0.25.1
pyee==11.1.0
pygments==2.17.2
pylibsrtp==0.10.0
Pympler==1.0.1
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyre-extensions==0.0.23
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291451001/work
python-dateutil==2.8.2
python-multipart==0.0.7
python-speech-features==0.6
python-version==0.0.2
pytorch-fid==0.3.0
pytorch-lightning==1.4.5
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
realesrgan==0.3.0
referencing==0.33.0
regex==2023.12.25
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
requests-oauthlib==1.3.1
retina-face==0.0.14
retinaface==1.1.1
rpds-py==0.17.1
rsa==4.9
ruamel-yaml-conda @ file:///home/conda/feedstock_root/build_artifacts/ruamel_yaml_1611943432947/work
safetensors==0.3.3
scenedetect==0.6.2
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
semantic-version==2.10.0
semver==3.0.2
sentry-sdk==1.40.5
setproctitle==1.3.3
six==1.15.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
starlette==0.36.3
streamlit==1.10.0
streamlit-webrtc==0.42.0
sympy==1.12
tb-nightly==2.16.0a20240121
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==1.1.0
test-tube==0.7.5
tf-keras-nightly==2.16.0.dev2023122410
threadpoolctl==3.2.0
tifffile==2023.12.9
timeout-decorator==0.5.0
tk==0.1.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchdata==0.7.1
torchmetrics==0.5.0
torchvision==0.15.2+cu117
tornado==6.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1621890532941/work
transformers==4.32.0
triton==2.0.0
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
tzlocal==5.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1657224465922/work
uvicorn==0.27.0.post1
validators==0.22.0
video2dataset==1.3.0
wandb==0.16.1
watchdog==4.0.0
wcwidth==0.2.13
webdataset==0.2.86
websockets==11.0.3
webvtt-py==0.4.6
werkzeug==3.0.1
wrapt==1.12.1
xformers==0.0.16
yapf==0.40.2
yarl==1.9.4
yt-dlp==2023.12.30
zipp==3.17.0
